+++ lib/fort.h
@@ -73,10 +73,33 @@ SOFTWARE.
 /*****************************************************************************
  *               RETURN CODES
  *****************************************************************************/
+
+/**
+ * Operation successfully ended.
+ */
 #define FT_SUCCESS        0
+
+/**
+ * Memory allocation failed.
+ */
 #define FT_MEMORY_ERROR   -1
+
+/**
+ * General error.
+ */
 #define FT_ERROR          -2
+
+/**
+ * Invalid argument.
+ */
 #define FT_EINVAL         -3
+
+/**
+ *  Libfort internal logic error.
+ */
+#define FT_INTERN_ERROR   -4
+
+
 #define FT_IS_SUCCESS(arg) ((arg) >= 0)
 #define FT_IS_ERROR(arg) ((arg) < 0)
 
@@ -936,6 +959,16 @@ int ft_set_cell_span(ft_table_t *table, size_t row, size_t col, size_t hor_span)
  */
 void ft_set_memory_funcs(void *(*f_malloc)(size_t size), void (*f_free)(void *ptr));
 
+
+/**
+ * Return string describing the `error_code`.
+ *
+ * @param error_code
+ *   Error code returned by the library.
+ * @return
+ *   String describing the error.
+ */
+const char *ft_strerror(int error_code);