+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - `ft_ln` returns status of operation.
 - Add new table property `adding_strategy` (2 strategies available - replace(default) and insert).
 - Add function `ft_row_count` (`row_count` in C++ API) to get number of rows in the table.
+- Add function `ft_strerror` to get string descriptions of error codes.
 
 ### Bug fixes